[fix] Reduced number of queries to database
The queries being made to database to differentiate friends
from the contacts list, was done in two to three steps, which included,
getting friends of users, comparing that with the contacts list, and in
some cases, like in adding or removing users, querying existing members.

Now, all of this is done in one query, where a list of pks is created
by filtering friends of users from contact list. If the length if this
list is equal to contacts list, the list is passed to add, remove, set
methods, which take care of duplicate keys internally